+++ providers/cryptowatch/markets.go
@@ -1,0 +1,11 @@
+package cryptowatch
+
+type Markets struct {
+	Result []struct {
+		ID       int    `json:"id"`
+		Exchange string `json:"exchange"`
+		Pair     string `json:"pair"`
+		Active   bool   `json:"active"`
+		Route    string `json:"route"`
+	} `json:"result"`
+}