+++ CHANGELOG.md
@@ -14,6 +14,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/) 
 1. Options for shorter output formats.
 1. New provider CoinGate.
 1. New option `--disable-pushover`.
+1. New option `--list-providers`.
 
 ## [0.2.0] - 2021-05-08